requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
Django==5.2.11
django-environ==0.12.0
django-filter==25.2
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.14.0
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.14
executing==2.2.0
idna==3.11
inflection==0.5.1
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==26.0
parso==0.8.4
parts==2.1.0
pexpect==4.9.0
pillow==11.2.1
products==2.0.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
smtp4dev==0.1.4
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.6.2
wcwidth==0.2.13
